+++ example/flutter/android/src/main/kotlin/com/example/flutterkmpexample/FlutterKmpExamplePlugin.kt
@@ -1,37 +1,16 @@
 package com.example.flutterkmpexample
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin
-import io.flutter.plugin.common.EventChannel
-import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
-import io.flutter.plugin.common.MethodChannel.MethodCallHandler
-import io.flutter.plugin.common.MethodChannel.Result
-import kotlinx.coroutines.GlobalScope
 
-/** FlutterKmpExamplePlugin */
-class FlutterKmpExamplePlugin: FlutterPlugin, MethodCallHandler {
-  /// The MethodChannel that will the communication between Flutter and native Android
-  ///
-  /// This local reference serves to register the plugin with the Flutter Engine and unregister it
-  /// when the Flutter Engine is detached from the Activity
-  private lateinit var channel : MethodChannel
+class FlutterKmpExamplePlugin: FlutterPlugin {
+  private val methodChannels = mutableListOf<MethodChannel>()
 
   override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
-    MyTestClassAndroid(flutterPluginBinding.binaryMessenger, GlobalScope)
-    
-    channel = MethodChannel(flutterPluginBinding.binaryMessenger, "flutter_kmp_example")
-    channel.setMethodCallHandler(this)
-  }
-
-  override fun onMethodCall(call: MethodCall, result: Result) {
-    if (call.method == "getPlatformVersion") {
-      result.success("Android ${android.os.Build.VERSION.RELEASE}")
-    } else {
-      result.notImplemented()
-    }
+    methodChannels += MyTestClassAndroid(flutterPluginBinding.binaryMessenger, SharedCoroutineScope)
   }
 
   override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
-    channel.setMethodCallHandler(null)
+    methodChannels.forEach { it.setMethodCallHandler(null) }
   }
 }