+++ flutter-kmp-stubs/src/main/java/io/flutter/plugin/common/MethodChannel.kt
@@ -1,6 +1,6 @@
 package io.flutter.plugin.common
 
-class MethodChannel(
+open class MethodChannel(
     messenger: BinaryMessenger,
     name: String,
 ) {