+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## unreleased
 
 - Update Kotlin to 2.1.0 and KSP to 2.0.21-1.0.25
+- Add support for iOS module generation
+- Make generated Android modules subclass MethodChannel to enable teardown
 
 ## v0.1.0-rc.0